Rework on_treeview_row_activated() logic, fixing a corner case
Previously, if a cell was empty, the next cell to the right was checked to
determine the action to be taken. This is counterintuitive when activating
the right-most column, as the check wraps around and the action is determined
by the left-most column instead of the expected middle column.

The now-unused _not_none() function was also removed.

svn path=/trunk/; revision=1229