Add support for svn-1.6.x and better detection of svn mv'ed files
svn-1.6.x changed the output of svn status, adding a column
and lines to describe tree conflicts.

http://subversion.tigris.org/svn_1.6_releasenotes.html

Any repository working copy used with svn 1.6.x
will be silently converted to a new format and not
usable any more with 1.5.x.

You'll have to make a fresh clone with 1.5.x if you
want to still use the old version.

This patch rework svn VC plugin so that both versions
1.5.x and 1.6.x are supported.

The code was parsing svn status output with fixed
string indexes, I changed that to use regexes that
work for the 2 svn versions.

I tested it working with 1.5.4 and 1.6.0 with:
- new files unknown to svn
- new files added
- modified files
- removed files
- moved files

svn path=/trunk/; revision=1334