Handle VC plugins not implemented methods by greying actions that use them.
The attached patch implements the greying of the
buttons attached to not implemented methods from
a VC plugin.

Crude, but working, I looked at inspect module, but
preferred the simple way of specifying the arguments.

Tested with RCS, SVN, git, and switching between
them to see button state toggling between sensitive
and not.

svn path=/trunk/; revision=1341