Create a CachedVc class to gather all common code duplicated between VC plugins that cache repository status informations.
There are slight behaviour changes:

bzr, monotone & git used to have a get_tree() which
always called lookup_tree() if it was not already cached,
but without caching at that time either. I think this was
wrong and only cut'n'pasted between plugins. But
would like to hear from the creators of those plugins.

The remaining of the patch is mostly renaming and
passing an additional "directory" argument to
_lookup_tree_cache() as was required for bzr, but
which remains unused for other plugins.