Make import problems easier to debug
Under windows we'll see far more import problems of
non-python-stdlib modules, for example the *gtk* family,
that has currently to be manually installed and that require
Path manipulations.

The attached patch makes the problem more easily
debuggable by reporting more precisely what is going on
with import failures.

I had the problem myself when trying meld on windows,
and Lorenz suggested to do the same thing when he tried:

http://mail.gnome.org/archives/meld-list/2009-April/msg00098.html