Improve blank line ignoring behaviour (closes bgo#505087 and bgo#392640)
Blank lines were not ignored when drawing areas in the textview or when
drawing regions in the diffmap, and blank regions were being incorrectly
skipped when navigating between diff chunks. This commit factors out the
correct logic (adapted from on_linkmap_expose_event()) and adapts it to
work with chunks returned by Differ.single_changes().

This commit fixes bgo#505087 and bgo#392640, as well as an unreported bug
with blank changes being shown in the diffmap.