Don't rely on HistoryFileEntry's file dialog existing
By just hiding the file dialog, we relied on the dialog not being deleted
somewhere else, such as in the default gtk.Dialog delete-event callback.
This led to a bug where closing a filechooser using the window manager
close button would cause a traceback when opening the dialog for a second
time. Destroying the dialog widget on response and setting our filechooser
member accordingly works around this problem.