Avoid missing pixels in linkmap insert/delete chunks (closes bgo#580311)
Painting of insert/delete chunks in the linkmap currently leaves a
single-pixel gap around the insertion point. This patch slightly improves
the visual appearance by having the lines meet up properly. The tradeoff
is that ending lines on the textview are now drawn one pixel higher, and
this makes changed chunks look not-quite-as-nice. However, it also avoids
a visual bug in three-way comparisons in the middle pane when a line gap
acts as both a chunk end and a chunk start.