Avoid retagging whole textbuffers if diff chunks haven't changed
Inline highlighting is a lot of work, so here we try as hard as possible
to avoid rehighlighting any chunks that have not changed or moved between
calls to update_highlighting(). While this still results in the occasional
excessive change computation, it massively speeds up the common case of
editing single lines.

This commit mostly fixes bgo#168139.

Based on a patch from Piotr Piastucki.