Add detailed cursor handling, and use for Up/Down sensitivity setting
There are several pieces of cursor-related information that we need: its
current location, which chunk (if any) it is currently in, and which chunks
are above and below it. The first is used for the status bar and for
determining the other two. The second may be used for a status-bar display
of where in the list of changes we currently are, and also for keyboard
operations (such as insert/delete/replace chunk). The third is useful in
skipping between changes, and in sensitivity setting.

This patch introduces a CursorDetails structure that stores all of the
interesting information, updates these details as necessary, and uses the
next/previous chunk information for sensitivity setting.

Differ: Add chunk by-line locator and by-index accessor for cursor handling
CursorDetails: New simple class holding cursor data
FileDiff: Use CursorDetails to keep track of cursor and its
          current/previous/next diff chunks
MeldApp: Use new next-diff-changed signal to update Up/Down sensitivity
MeldDoc: Add new cursor-related signals

Closes bgo#533752 and bgo#609629