Move ignore-blank-lines implementation to Differ class
One problem with the ignore-blank-lines preference is that the current
implementation causes FileDiff to use a different set of diff chunks to
those provided by Differ. Moving the implementation of the preference into
the Differ class unifies these views, and makes it easier to extend the
information provided by Differ.

Also, the new implementation in Differ._consume_blank_lines() is superior
in that it rewrites diff chunks whose nature changes when blank lines are
ignored. For example, a 'replace' chunk in which one side consists entirely
of blanks lines is now correctly transformed into an 'insert' or 'delete'
chunk, as appropriate.