Bring back merge arrows in conflict blocks (closes bgo#602871)
In the 1.2.x series, Meld had arrows for conflict blocks. However, with
recent improvements to the conflict detection and separation, these arrows
got lost, and no longer made sense. For example, with the new conflict
detection, we can have one block of a conflict be empty; it makes no sense
to use a normal 'replace' arrow for an empty block.

This commit passes around some extra information to enable the drawing and
processing to do sensible things for conflict blocks.