Update git to use "git diff-index" instead of "git status"
"git status" is a "porcelain" command whose output is shown in a format
that is easy for a user to digest.  The output formatting can change as
it is not considered a "plumbing" git command.

By using "git diff-index" instead of "git status" we reduce the chance
that meld will mis-parse a git tree's current status if/when the output
of "git status" changes.  Additionally, "git diff-index"'s output is
much easier to parse which simplifies meld's code a bit.

Lastly, unlike "git status", "git diff-index" can be used to only
show the changed files in a specific repository location.  "git status"
is only able to show all changed files in a repository.  This feature
of "diff-index" will be useful when updating meld to properly parse
git repositories' sub-directories. (see bug 609025)

Signed-off-by: Peter Tyser <ptyser@gmail.com>