Split out DiffMap functionality into new widget
The DiffMap widget (the bars showing a coloured summary of changes at
the left and right of the main file comparison window) has always been
a purely custom creation based on gtk.DrawingArea. This commit breaks it
out into a separate file and makes it a DrawingArea subclass.

This also changes the drawing done, syncing the colours used with the
LinkMap and TextView, and allowing thinner lines for insert chunks. In
addition, our style-dependent calculations should now be more robust for
unusual themes.