Rework undo behaviour; support undo/redo after save (closes bgo#156984)
The current undo stack implementation clears all undo information on
save, because its modification tracking doesn't allow for easily
repositioning the modification point within the stack. This commit adds
per-buffer tracking of modification points within the stack, and
repositions these as appropriate on save. In addition, we now emit
signals when the modified state of a buffer (as defined by the undo
stack) changes.