write_pipe(): Discard stdout output
Previously, meld would print any stdout output from a command executed
via the write_pipe() function to the command line.  As an example:
  $ ./bin/meld ./
  (Double click on the git.py file in the meld GUI)
  patching file meld/vc/git.py
  (Double click on the _vc.py file in the meld GUI)
  patching file meld/vc/_vc.py

The "patching file XYZ" output from the 'patch' command litters the
terminal window used to execute meld and doesn't provide any useful
information.

This change discards the stdout output of a command executed via the
write_pipe() function.  Note that stderr output will still be
printed to the command line as it can be useful in diagnosing why
a sub-command is having errors.

Signed-off-by: Peter Tyser <ptyser@gmail.com>