Fix behavior when "Save modified files?" window is closed
After modifying a file, not saving the changes, and attempting to close
the file or quit meld, a user is presented with a "Save modified
files?" dialog window.  Previously, pressing the escape key at this window
or closing the window would result in the modified file being closed and
the modifications lost.  This is at odds with most other applications
which assume when a user presses the escape key or closes the dialog
window they only want to close the dialog window, not discard their
modifications.  This change makes meld behave similar to other
applications.

Signed-off-by: Peter Tyser <ptyser@gmail.com>