Remove option to edit files in Meld and simplify preference interface
We currently provide options to edit files in Meld, the system editor,
or an external editor. However, Meld is not designed as a stand-alone
editor; the system editor is almost certain to be a superior editing
tool. This commit removes the option to edit files in Meld, leaving
the system editor and custom editor options.

This only changes the behaviour of selecting "Open" from the context
menu in various views. Starting a 'comparison', even if on a single
file, will always open in Meld.