git: Update diff-logic
Previously, meld would use the 'git diff-index HEAD' command to find out
the status of modified files.  This command would compare the HEAD revision
to the files on disk.  This command did not take into consideration git's
"index" which resulted in odd behavior such as not handling unmerged files
correctly as pointed out in https://bugzilla.gnome.org/show_bug.cgi?id=617098

Separating 'git diff-index HEAD' into 'git diff-index --cached HEAD' and
'git diff-files' proprerly accounts for git's "index".  Now modifications
to files on disk, in the index, or unmerged files are shown in meld.

Note that the current functionality isn't perfect as a user can't decide
if they do/don't want to take git's "index" into consideration for
their diffs.  Ideally a feature such as an "index filter" will be
added in the future to allow finer control over if and how the contents
of the "index" are shown.

Signed-off-by: Peter Tyser <ptyser@gmail.com>