Fix undo for grouped actions
The recent checkpoint-related undo rework requires every action to have
an associated buffer. While this is true of all of our basic actions, it
breaks for grouped actions. This commit adds a slightly hacky workaround
by assuming that a group of actions all operates on the same buffer.
While this is currently accurate, future changes could break it.