Highlight currently focused change
Now that we have keyboard actions on the currently selected commit, we
also need to show which change Meld thinks the cursor is in. This patch
indicates the current change by adding a highlight to the linkmap and
textview fills.

Note that ideally we shouldn't queue_draw everything when the focused
chunk changes.