vc: Add ability to check if version control repos are valid
Sometimes meld may be ran in a directory that contains invalid or
corrupt version control information.  (eg run 'mkdir .svn; meld .' in a
git repository).  Previously meld would not gracefully handle invalid
repositories.  This patch attempts to add a quick sanity check
to most version control systems.  The checks are meant to be
as quick as possible to reduce startup time.

Signed-off-by: Peter Tyser <ptyser@gmail.com>