Allow reordering of tabs
This can be useful when many diffs are open in tabs.  For example,
say diffA, diffB, and diffC are open in tabs, and diffA and diffC
have similar changes.  To quickly compare the changes in diffA
to diffC it is convenient to have the 2 tabs next to each other
so that they can quickly be toggled between.  Allowing reordering
lets the user move the diffA tab to be next to the diffC tab for
quick comparison.

Signed-off-by: Peter Tyser <ptyser@gmail.com>