Add ability to set the tooltip text of a tab
Previously, the tab title and tab tooltip text had to be identical.
Allowing different titles and tooltips can be useful, so add the
ability to set them individually.

The one functional change this patch adds is that the tooltip for a
directory or VC diff tab now says "Location: <diff directory location>".
Previously it just said the basename of directory the diff was occuring
in, which wasn't very descriptive.

Signed-off-by: Peter Tyser <ptyser@gmail.com>