vcview: Make temporary version control files un-writable
A version control diff is a diff of a local file vs the contents
of a repository.  A user can't edit the repository directly, so they
shouldn't be able to edit the temporary file representing the repository
either.  This also prevents a user from accidentally modifying a
temporary version control file in /tmp, saving it (thinking they were
saving the "real", working copy), then having it be automatically
removed when closing meld.

Signed-off-by: Peter Tyser <ptyser@gmail.com>