Rework VC and DirDiff TreeViews to take text attributes from the store
The existing code used string substitution into markup to format
cells in the TreeViews. While this worked, the approach was messy to
theme. With this commit, we instead store pango, etc. attributes in
the TreeStore and set column properties appropriately.

While this should make themeing improvements easier, there are no
themeing changes in this commit.