Correctly handle unicode linebreaks (closes bgo#602933, bgo#627940)
As it stands, Meld relies on Python's universal newline support to
translate various traditional newline sequences into '\n', which we
then split on. This approach fails for unicode linebreaks, and when
Python's universal newline support isn't available.

This patch moves to using a combination of Python's unicode
splitlines() and asking our gtk.TextBuffer where it thinks linebreaks
are in order to handle unicode linebreaks.