Add "Copy change [left/right] [above/below]" actions (closes bgo#606613)
These actions add the ability to copy chunks in FileDiff from the
menus, based on the currenly focused change. These are the final
actions in the set of things that could always be done via the central
diff bar, but weren't available for keyboard navigation. Now, they're
available in menus, which makes them significantly more discoverable.