Rework directory-level file comparison (closes bgo#586656)
In comparison to previous incarnation, this reworked implementation
features somewhat simpler logic, progressive file reading, binary file
filter handling (i.e., we now disable filters if we guess that a file
is binary) and provides earlier exit in no-filter cases.

We also define and use different exit codes from the function; we can
now additionally distinguish files that appear to be the
same/different, but for which a full comparison was not run, and error
cases.

This commit also adds the Python 2.4 compatible namedtuple from:
    http://code.activestate.com/recipes/500261/
as a replacement for misc.struct.