CachedVC: Add ability to update file status when its modified
Previously, a VcView diff of a cached version controlled directory could
easily become out of date.  For example, if the file 'AUTHORS' was a
version controlled file in a git repository:
  > echo "asdf" >> AUTHORS
  > meld ./
  # Double click on the AUTHORS file to view its diff.
  # Remove the "asdf" at the bottom of a file and save it.
  # Close the AUTHORS tab.
  # The state of the AUTHORS file will still show up as modified, even
  # though it is now unmodified.

This change implements the new update_file_state() for git.  Similar
funcationality could be added to monotone, darcs, bzr, and arch.

Signed-off-by: Peter Tyser <ptyser@gmail.com>