Add a toggle for unlocking scrolling in file diff (closes bgo#348119)
While it's expected that the normal mode of operation in Meld is to
have panes' scrolling locked, it is sometimes useful to be able to
unlock scrolling to compare two different sections of files.

This commit adds a "Lock scrolling" option to the View menu of file
comparisons, and adds a toggleable lock button to the upper-right
corner of comparison tabs.

The locked and unlocked icons are taken from the GNOME Colors
project, licensed under GPL v2.