When checking for bzr repo, only use quick checks (closes bgo#641624)
'bzr check' checks the working tree, branch and repository. Since we're
only running this to verify that the directory is in fact a bzr repo,
we don't really need the full suite of checks. This commit changes to
just performing tree and branch checks, for a considerable speedup on
large repositories.

Patch from Tristan Hill.