Add action to cycle between panes in FileDiff (closes bgo#626186)
One of the last pieces for allowing keyboard-only merging is being able
to cycle between panes. This commit adds an action to the View menu
that allows this.

The tricky decision to be made in this feature is where to place the
cursor when changing panes. The previous cursor location will often be
irrelevant to the user's current action, and is a bad default. In the
common case, we simply guess which line in the moved-to chunk
corresponds to the current line in the current chunk.

In order to avoid disturbing in-progress editing that may be going on,
we leave the cursor in place if it is already in the moved-to chunk.