Check whether the inserted bit is a sole CR (closes bgo#650726)
If a file ends in a sole CR, the handling for inserting split CR/LF
pairs results in an infinite loop. This commit fixes this by checking
to see whether we're inserting a single CR; this should only happen
at the end of a file, and so shouldn't affect other cases.