Don't split lines on formfeed characters (closes bgo#652996)
This patch resolves a disagreement between GtkTextStore and Python's
unicode splitlines() about what constitutes a line break. Specifically,
Python's splitlines() will break on a formfeed character, and
GtkTextStore will not. We work around this by scanning line ends in our
split strings for FF characters and re-joining lines as appropriate.

This problem only exists in Python 2.7 or later; 2.6 and earlier are
unaffected. The bug describing the relevant Python change is:
  http://bugs.python.org/issue7643