Handle metadata-only changes in Git (closes bgo#655315)
In situations where a file has a metadata-only change (e.g., mode
changes or empty file creation), Git's diff consists of an empty
patch preceeded by several lines of metadata. In the end, this ends
up as an invalid patch that won't apply, leading to the dreaded
Patch Failed Box of Doom.

This commit adds pruning of metadata from git-generated diffs, makes
sure that an empty-of-content git diff results in an actually empty
patch, and changes our git plumbing queries to parse out permission
change information so that we can show it in the UI.

With this change, we're abusing the "Options" column in the VC UI to
show essentially random extra information. However, reworking that UI
so that it makes sense across VCs is a huge project, and Git doesn't
otherwise use the "Options" column.