Add new VC file request interface and SVN support
Currently, all VC plugins are asked for a diff between the on-disk
contents and the repository head, which VCView then reverse applies
with patch to get a temporary copy of the repository contents. This
works surprisingly well, but has several issues, most notable of which
is that it's borderline impossible to figure out what's happening when
it breaks. In addition, some VCSs (e.g., subversion) give out patches
that won't actually reverse-apply because of other magic that's been
performed on the checkout, but isn't reflected in the diff.

This patch adds a simple 'give me the file' API to the Vc class, and
adds support for it to the svn plugin. While it has a slot for a
commit argument to retrieve non-HEAD commits, this currently doesn't
work and has no UI anyway.