Assume same file type if not available for all files (closes bgo#658407)
For the purpose of syntax highlighting, we always try to determine the
mime type of files, currently using only the filename. In situations
where the filename tells us nothing, we can reasonably assume that all
files have the same type.

This commit changes our behaviour so that, if we get one useful file
type, we use it to fill in missing file types in the remaining files.
This should specifically help with cases where temporary files are
given different extensions. e.g., TortoiseSVN uses .mine, .working, etc.