Prompt user for action when file changes externally (closes bgo#460201)
When we get a notification that a file has changed externally, we now
check mtimes, maintaining both the last mtime we saw (to prevent
duplicate notifications) and the mtime at which we loaded the file.

This doesn't yet prompt the user to overwrite a changed file; they've
already been prompted to reload, but we should still prompt for the
overwrite.