Add and use sorting criteria for VC systems (closes bgo#663055)
This fixes the problem of not knowing which VC the user wants to see
when a directory is under revision control by multiple different
systems.

The sort order is in rough order of popularity, with distributed
version control systems coming first. The reasoning for this is that
if we find e.g., SVN and Git in a single directory, the likelihood is
that there is an organisation SVN repository and the Git copy is the
user's local working repository. Therefore, we show the Git copy.