Handle activating fake lines in directory comparison
We get fake lines in DirDiff when we have empty folders. This commit
introduces some sanity checking on path usage so that we don't get a
traceback when someone double-clicks one of these empty placeholder
rows.

In this case, 'handle' means 'do nothing', since there is no sensible
action to take.