Add and use New Comparison tab, replacing the existing New dialog
The New Comparison tab provides a clear workflow that users can follow
to start a comparison, hopefully without too much friction for power
users.

The more permissive approach of just providing three filechoosers and
working things out from there might be preferrable, but shortcomings
with GtkFileChooser (i.e., you can't select both files and directories
from the same chooser) mean that it's not feasible.

This implementation avoids one of the pitfalls of the existing dialog
by requiring the user to click on file/folder/vc before browsing for
files. While this does add an extra click in the file comparison case,
it avoids having the user start to browse for a folder in a file
selection dialog.

In order to help users who just want a blank comparison, there is a
button that opens a new tab of the selected type with no
files/folders/etc. selected. Note that this currently doesn't work as
intended for VC and directory comparisons.

In addition, if we're not given anything to compare on startup, a
New Comparison tab is now opened.

Closes bgo#342701, bgo#349926, bgo#448658, and bgo#518179.