Fix sensitivity of prev/next diff in VC and DirDiff (closes bgo#650181)
Common code for finding the previous/next differing rows was moved to
the tree module, and both VCView and DirDiff now use that approach.
Previous/next diff is updated on tree cursor movement and used to set
sensitivity of the corresponding actions.

Currently there is no attempt to minimise the number of tree
traversals done when the cursor moves. In some situations of very large
trees with few changes, this may mean that scrolling through the tree
with the keyboard will be very slow. There are certainly tricks that
can be done with invalidation of the prev/next validity (TODO-ed in the
code) that should improve the situation.