Support new tab opening using D-Bus (closes bgo#616466, bgo#453670)
The goal of this patch is to support opening new tabs in an existing
instance from the command line, enabled by passing a new command line
flag. Communication between instances is enabled by D-Bus.

New tab opening is not the default behaviour, since many VC systems
launch Meld and block waiting for an exit code to indicate success or
failure when doing merges, and we can't break this expectation.

This commit introduces an optional dependency on dbus packages.

Based on a patch from Kacper Wysocki and Antoine.