Cull chunk linking visualisation in LinkMap to decrease visual clutter
In Meld's view of chunks, it's very easy to have an on-screen chunk in
one pane whose corresponding chunk in another pane is completely
off-screen. This doesn't look great, because the LinkMap curve joining
these two chunks curves up (or down) into oblivion, and doesn't really
convey any useful information, while cluttering the display. In
addition, we allow users to use the normal LinkMap actions on these
chunks, even though it's impossible to see clearly what the effects
will be.

This change adds culling of chunk display, such that the normal linked
visualisation between chunks will only be used if at least some part
of each chunk is on the screen. Otherwise, a simple curved cap will be
shown, and the LinkMap actions usually associated with that chunk will
be hidden and disabled.