Draw currently displayed page area in DiffMap (closes bgo#470729)
While ideally the scrollbar thumb should give a decent indication of
the location and size of the current viewing pane in terms of the
whole file, it doesn't always correspond very well. For example, in
long files, the scroll thumb will not shrink below a certain size for
usability purposes, which damages its use as an indicator.

This commit introduces a relatively subtle overlay on our DiffMaps to
show which part of the file is currently being displayed.