Fix icon colour tinting, and add luminance adjustments
This commit fixes a bug where we didn't convert integer RGB values to
Cairo floats before setting colours, and also introduces an adjustment
for the perceived brightness of a tint colour. When adding a tint to an
icon, we get very different resulting brightness depending on what tint
colour was added. This adds some simple heuristics to try and determine
the perceived luminance of the tint, and adjusts accordingly.