Improve popup right-click behaviour in tree views (closes bgo#669828)
Existing behaviour when right-clicking on an unselected row was to not
select the row, but show a context menu for the other currently
selected rows, which is inconsistent with Nautilus, among others.

This patch changes our behaviour to be more consistent with other file
management tools, in both Meld's directory and VC comparison views.
Right clicking an unselected file now selects it, moves the cursor to
it, and opens the context menu. The existing right-click behaviour for
already-selected files should be preserved.