Fix refresh action to not reset modified flags (closes bgo#670133)
Our existing "Refresh" action piggy-backed on the reload code, using
a semantically-dubious empty list to indicate a lack of changes. One
upshot of this was that undo-related functionality like modified flags
were unintentionally cleared on refresh.

This patch breaks out minimal required refresh functionality into a
new method, and replaces the old set-empty-files idiom with a call to
this method.