Clean up FindBar table and work around gtk.Table row spacing brokenness
In a gtk.Table, row_spacing is displayed regardless of whether a row is
visible or not. This was affecting FindBar when the Replace row was not
shown. This commit cleans up the FindBar table somewhat and adds a hack
to get rid of the undesired row spacing in Find mode.