Inline highlighting performance improvements
The default preprocessing in MyersSequenceMatcher removes all items
that do not appear in the other file. It works pretty well for lines
which are long and unique. However, it fails for character-based
matching, as both files usually use the same alphabet.

This patch adds a new matcher class called InlineMyersSequenceMatcher
which provides additional optimization based on the assumption that a
single character-based chunk must be at least 3 characters long.