Don't try to open non-existent paths from directory comparison
The "Open externally" command always tried to open file paths gathered
from the model, regardless of whether any paths were gathered. This led
to attempts to open None in cases where the tree contained errors or
missing files. This commit fixes this by first sanitising the list of
files to be opened.