Rework external opening to better obey file types
Our existing code had several issues, most obvious of which was that
it would try to open non-text files in a text editor when there was a
custom editor set.

This patch fixes _open_files to launch the correct application more
reliably, using gio's file type functionality. In theory this is
cross-platform, though in reality file sniffing on other systems
appears less-than-awesome. Since the fallback goes to generic OS
'open this file' calls in each case, this shouldn't be a big problem.