Identify changed chunks in comparison updates and use when highlighting
With this change, we now identify which chunks in an existing
comparison are changed by a comparison update. This information is
then passed on to the inline highlighting updater and used to only
re-highlight changed regions of our text buffers.

As a result of these changes, we now also avoid manually updating
highlighting and instead simply perform highlighting in a callback.