Fix regression in three-way diffs and on-the-fly updates
This commit fixes a regression in commit 92295b5 that caused three-way
diffs and on-the-fly updating to fail because they were synthesizing
new chunks, but not using the namedtuple representation.

Included in this fix is a small rework of some of the code in
Differ._change_sequence, being more explicit about what the various
loops are actually doing.