Try to restore a valid tree cursor on row deletion (closes bgo#680571)
When removing rows from directory treeviews (e.g., by hiding or
deletion) we could easily end up with current_path pointing to a
now-invalid tree path, causing tracebacks.

Rather than simply invalidating our current_path, this patch actually
tries to move the tree cursor to the 'next' valid location, in much the
same way as traditional file managers will when a file is deleted. The
logic is roughly that we try to select the next successor if there is
one, otherwise the next predecessor, otherwise the nearest parent. As a
fallback, the current_path is simply invalidated.