Disconnect filter signals for removed tabs (closes bgo#681422)
When removing a tab, we weren't disconnecting the various 'filter
changed' signals on the app. With these tabs not properly unrefed, the
signals would continue to fire after the widget tree had already been
destroyed, causing tracebacks and badness.

This patch adds cleanup of these app-level signals to the delete events
of individual DirDiff and FileDiff tabs.