Fix files labelled as empty folders when deleting (closes bgo#684350)
When recursively deleting a folder, we recursively update the tree row
in the scheduler, but then remove the row immediately. This leads to
a recursive update being scheduled on the row that replaces the deleted
row.

While this ordering is very broken, the risks involved in messing with
the scheduled updates and deletion ordering are high, given that there
are no other functional bugs reported against this. As such, this
commit simply adds a specific check for recursive updating of a file
row, avoiding the 'empty folder' appearance.