Check for modifier keysyms only on first level (closes bgo#683534)
There is a problem when checking for keypress/release events on
modifier keys as they do not necessarily have the modifier keysym
on that level as expected (see bugs #683534 and #584342).
By checking for the keysym on the first level (and thereby assuming
a modifier key has its appropriate keysym set on the first level)
we can gracefully circumvent this problem.

This might not be the ideal solution, but considering there is no
good way to retrieve the current modifier state in PyGTK it is at
least better than previously.
The ugly workaround for #584342 is thereby obsolete.

We use translate_keyboard_state method instead of lookup_key, as
the latter one returns wrong values in some cases with multiple
layouts.